feat(cli): move into internal folder
