+++ internal/cmd/eval.go
