+++ main.go
@@ -1,6 +1,6 @@
 package main
 
-import "github.com/operator-framework/combo/cmd"
+import "github.com/operator-framework/combo/internal/cmd"
 
 func main() {
 	cmd.Execute()
